pyproject.toml
[project]
name = "personal-clone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agent-starter-pack>=0.29.3",
    "aiosqlite>=0.22.1",
    "apscheduler>=3.11.2",
    "bs4>=0.0.2",
    "fastapi>=0.123.10",
    "google-adk>=1.21.0",
    "google-api-python-client>=2.187.0",
    "google-cloud-aiplatform<1.132.0",
    "google-cloud-bigquery-storage>=2.36.0",
    "google-genai>=1.56.0",
    "httpx>=0.28.1",
    "isodate>=0.7.2",
    "jinja2>=3.1.6",
    "keepa>=1.4.3",
    "litellm>=1.80.11",
    "lxml>=6.0.2",
    "pinecone[asyncio]>=8.0.0",
    "pydantic>=2.12.5",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
    "rich>=14.3.3",
    "slack-bolt>=1.27.0",
    "uvicorn>=0.41.0",
    "youtube-transcript-api>=1.2.3",
]

[dependency-groups]
dev = [
    "ipython>=9.10.0",
    "pyperclip>=1.11.0",
]
